+++ dffml/source/hdfs_source.py
@@ -16,7 +16,7 @@ class HDFSSourceConfig(BaseConfig, NamedTuple):
     port: str
     user: str
     filepath: str
-    source: str
+    source: BaseSource.load
 
 
 @entry_point("dffml.source")
@@ -26,15 +26,12 @@ class HDFSSource(BaseSource):
 
     async def __aenter__(self) -> "BaseSource":
         self.client = InsecureClient(
-            "http://"+self.config.host+":"+self.config.port,
+            "http://" + self.config.host + ":" + self.config.port,
             user="hadoopuser",
         )
-        with self.client.read(self.config.filepath, encoding="utf-8", delimiter="\n") as reader:
-           pass
+        with self.client.read(self.config.filepath, encoding="utf-8") as fd:
+            await self.config.source.load_fd(fd)
         return self
-
-    async def __aexit__(self, exc_type, exc_value, traceback):
-        pass
 
     @classmethod
     def args(cls, args, *above) -> Dict[str, Arg]:
@@ -42,15 +39,20 @@ class HDFSSource(BaseSource):
         cls.config_set(args, above, "port", Arg(type=str))
         cls.config_set(args, above, "user", Arg(type=str))
         cls.config_set(args, above, "filepath", Arg(type=str))
-        cls.config_set(args, above, "source", Arg(type=str))
+        cls.config_set(args, above, "source", Arg(type=BaseSource.load))
+        for loaded in BaseSource.load():
+            loaded.args(args, *cls.add_orig_label(*above))
         return args
 
     @classmethod
     def config(cls, config, *above):
+        source = cls.config_get(config, above, "source")
+        source = source.withconfig(config, *cls.add_label(*above))
+        source._open = cls.__aenter__.__get__(source)
         return HDFSSourceConfig(
             host=cls.config_get(config, above, "host"),
             port=cls.config_get(config, above, "port"),
             user=cls.config_get(config, above, "user"),
             filepath=cls.config_get(config, above, "filepath"),
-            source=cls.config_get(config, above, "source"),
+            source=source,
         )